+++ Projects/Shared/Utility/Project.swift
@@ -1,0 +1,28 @@
+//
+//  Project.swift
+//  Manifests
+//
+//  Created by 조용인 on 6/16/25.
+//
+
+import ProjectDescription
+import ProjectDescriptionHelpers
+
+let project = Project(
+  name: "Utility",
+  organizationName: "Sseudam.a2bo.ios",
+  options: .default,
+  settings: .default,
+  targets: [
+    .target(
+      name: "Utility",
+      destinations: .iOS,
+      product: .staticLibrary,
+      bundleId: "Sseudam.a2bo.ios.Utility",
+      deploymentTargets: .iOS("17.0"),
+      infoPlist: .extendingDefault(with: [:]),
+      sources: ["Sources/**"]
+    )
+  ]
+)
+
